chore: clear technical, product, and architectural debt
Addresses the open debt tickets on this branch; backend and frontend
suites are green (991 / 854).

- #227 [Security] require auth on 4 projectKey reads (cross-tenant read)
- #224 fix 6 discarded-BadRequest guards in KeyController
- #229 DeleteAsysnc -> DeleteAsync across the interface + call sites
- #230 ConfigConroller.cs -> ConfigController.cs; canonical Cloud/Clouds
       routes added, old kept [Obsolete] delegating
- #231 move LocalizationSecret out of the Blocks.Genesis namespace
- #233 add POST alongside DELETE for DeleteKeys (proxies drop DELETE body)
- #234 client kebab-case filenames; centralize hardcoded gateway paths
- #235 gitignore + untrack build/scratch artifacts (incl. a 29.7 MB db)
- #236 naming enforcement: .editorconfig + CONTRIBUTING.md

Product naming standardized to "Blocks Localization" on the README. Renames
follow obsolete-old-keep-new: the old symbol, route, or wire field is kept and
marked [Obsolete] / @deprecated, forwarding to the new name. The delete-module
feature (#228, intentionally disabled) and grant-breaking scope fixes (#232)
are deferred with reasons on their tickets.
