+++ server/XUnitTest/Api/Controllers/KeyControllerTests.cs
@@ -498,7 +498,7 @@ namespace XUnitTest
             var expectedResponse = new BaseMutationResponse { IsSuccess = true };
 
             _keyManagementServiceMock
-                .Setup(x => x.DeleteAsysnc(request))
+                .Setup(x => x.DeleteAsync(request))
                 .ReturnsAsync(expectedResponse);
 
             // Act
@@ -542,7 +542,7 @@ namespace XUnitTest
             var failureResponse = new BaseMutationResponse { IsSuccess = false };
 
             _keyManagementServiceMock
-                .Setup(x => x.DeleteAsysnc(request))
+                .Setup(x => x.DeleteAsync(request))
                 .ReturnsAsync(failureResponse);
 
             // Act
@@ -955,9 +955,10 @@ namespace XUnitTest
         }
 
         [Fact]
-        public async Task Delete_WithNullRequest_ThrowsNullReferenceException()
+        public async Task Delete_WithNullRequest_ReturnsBadRequest()
         {
-            await Assert.ThrowsAsync<NullReferenceException>(() => _controller.Delete(null));
+            var result = await _controller.Delete(null);
+            result.Should().BeOfType<BadRequestObjectResult>();
         }
 
         [Fact]
@@ -977,12 +978,12 @@ namespace XUnitTest
         }
 
         [Fact]
-        public async Task TranslateAll_WithNullRequest_ReturnsOk()
+        public async Task TranslateAll_WithNullRequest_ReturnsBadRequest()
         {
-            // BadRequest(...) is discarded; the null request flows to the service and Ok is returned.
+            // Null request is now short-circuited with BadRequest instead of flowing to the service.
             var result = await _controller.TranslateAll(null);
 
-            result.Should().BeOfType<OkObjectResult>();
+            result.Should().BeOfType<BadRequestObjectResult>();
         }
 
         [Fact]
