+++ server/XUnitTest/Api/ServiceRegistryTests.cs
@@ -8,6 +8,7 @@ using Storage.DomainService.Storage;
 using Storage.DomainService.Storage.Validators;
 using Worker;
 using Eurolm.DomainService.Shared.Entities;
+using Eurolm.DomainService.Shared.Utilities;
 using Eurolm.DomainService.Services;
 using Eurolm.DomainService.Repositories;
 using Eurolm.DomainService.Services.HelperService;
