+++ server/XUnitTest/Services/Language/LanguageManagementServiceTests.cs
@@ -197,7 +197,7 @@ namespace XUnitTest
                 .Returns(Task.CompletedTask);
 
             // Act
-            var result = await _service.DeleteAsysnc(request);
+            var result = await _service.DeleteAsync(request);
 
             // Assert
             result.Should().NotBeNull();
@@ -215,7 +215,7 @@ namespace XUnitTest
                 .ReturnsAsync((BlocksLanguage)null);
 
             // Act
-            var result = await _service.DeleteAsysnc(request);
+            var result = await _service.DeleteAsync(request);
 
             // Assert
             result.Should().NotBeNull();
