--- server/XUnitTest/XUnitTest.csproj.Backup.tmp
@@ -1,30 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-   
-    <IsPackable>false</IsPackable>
-    <IsTestProject>true</IsTestProject>
-    <SonarQubeExclude>true</SonarQubeExclude>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="coverlet.collector" />
-    <PackageReference Include="JunitXml.TestLogger" />
-    <PackageReference Include="FluentAssertions" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" />
-    <PackageReference Include="Moq" />
-    <PackageReference Include="xunit" />
-    <PackageReference Include="xunit.runner.visualstudio" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Api\Api.csproj" />
-    <ProjectReference Include="..\DomainService\Eurolm.DomainService.csproj" />
-    <ProjectReference Include="..\Worker\Worker.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Using Include="Xunit" />
-  </ItemGroup>
-
-</Project>
